Les formats sont les mêmes peu importe la langue.
